Refactor: move Element_Basis_Index::construct_range() to module_ao
