+++ source/source_lcao/module_ri/ABFs_Construct-PCA.cpp
@@ -2,7 +2,7 @@
 
 #include "../../source_base/module_external/lapack_connector.h"
 #include "../../source_base/global_function.h"
-#include "../../source_base/element_basis_index.h"
+#include "../../source_basis/module_ao/element_basis_index-ORB.h"
 #include "../../source_base/matrix.h"
 #include "../../source_lcao/module_ri/Matrix_Orbs11.h"
 #include "../../source_lcao/module_ri/Matrix_Orbs21.h"
