+++ source/source_lcao/module_ri/LRI_CV.hpp
@@ -10,6 +10,7 @@
 #include "LRI_CV_Tools.h"
 #include "exx_abfs-construct_orbs.h"
 #include "RI_Util.h"
+#include "../../source_basis/module_ao/element_basis_index-ORB.h"
 #include "../../source_base/tool_title.h"
 #include "../../source_base/timer.h"
 #include "../../source_pw/module_pwdft/global.h"
